requirements.txt
aiodns==3.1.0
aiofiles==23.1.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
frozenlist==1.4.0
idna==3.4
loguru==0.7.2
lxml==4.9.3
magic-filter==1.0.12
multidict==6.0.4
origamibot==2.3.6
pycares==4.4.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
typing_extensions==4.7.1
unicodedata2==15.1.0
urllib3==2.0.6
win32-setctime==1.1.0
yarl==1.9.2
